updates Account Page file structure
