--- resources/assets/components/pages/AccountPage/BadgeImages.js
@@ -1,43 +1,0 @@
-import signupBadge from './Badges/Signup.svg';
-import signupBadgeLocked from './Badges/SignupBadgeLocked.svg';
-import onePostBadge from './Badges/Posts1.svg';
-import onePostBadgeLocked from './Badges/Posts1Locked.svg';
-import twoPostsBadge from './Badges/Posts2.svg';
-import twoPostsBadgeLocked from './Badges/Posts2Locked.svg';
-import threePostsBadge from './Badges/Posts3.svg';
-import threePostsBadgeLocked from './Badges/Posts3Locked.svg';
-import breakdownBadge from './Badges/BreakdownBadge.svg';
-import breakdownBadgeLocked from './Badges/BreakdownBadgeLocked.svg';
-import oneStaffFaveBadge from './Badges/StaffFave1.svg';
-import oneStaffFaveBadgeLocked from './Badges/StaffFave1Locked.svg';
-import twoStaffFavesBadge from './Badges/StaffFave2.svg';
-import twoStaffFavesBadgeLocked from './Badges/StaffFave2Locked.svg';
-import threeStaffFavesBadge from './Badges/StaffFave3.svg';
-import threeStaffFavesBadgeLocked from './Badges/StaffFave3Locked.svg';
-import voterBadge from './Badges/VoterBadge.svg';
-import voterBadgeLocked from './Badges/VoterBadgeLocked.svg';
-import lock from './Badges/Lock.svg';
-
-const BadgeImages = {
-  signupBadge,
-  signupBadgeLocked,
-  onePostBadge,
-  onePostBadgeLocked,
-  twoPostsBadge,
-  twoPostsBadgeLocked,
-  threePostsBadge,
-  threePostsBadgeLocked,
-  oneStaffFaveBadge,
-  oneStaffFaveBadgeLocked,
-  twoStaffFavesBadge,
-  twoStaffFavesBadgeLocked,
-  threeStaffFavesBadge,
-  threeStaffFavesBadgeLocked,
-  voterBadge,
-  voterBadgeLocked,
-  breakdownBadge,
-  breakdownBadgeLocked,
-  lock,
-};
-
-export default BadgeImages;
