+++ resources/assets/components/pages/AccountPage/Badges/Badge.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import badgeImages from './BadgeImages';
-import { BaseFigure } from '../../utilities/Figure/Figure';
+import { BaseFigure } from '../../../utilities/Figure/Figure';
 
 const Badge = props => {
   const { name, earned, size, className, children, showLock } = props;
