+++ resources/assets/components/pages/AccountPage/Badges/BadgeImages.js
@@ -1,0 +1,43 @@
+import signupBadge from './BadgesImages/Signup.svg';
+import signupBadgeLocked from './BadgesImages/SignupBadgeLocked.svg';
+import onePostBadge from './BadgesImages/Posts1.svg';
+import onePostBadgeLocked from './BadgesImages/Posts1Locked.svg';
+import twoPostsBadge from './BadgesImages/Posts2.svg';
+import twoPostsBadgeLocked from './BadgesImages/Posts2Locked.svg';
+import threePostsBadge from './BadgesImages/Posts3.svg';
+import threePostsBadgeLocked from './BadgesImages/Posts3Locked.svg';
+import breakdownBadge from './BadgesImages/BreakdownBadge.svg';
+import breakdownBadgeLocked from './BadgesImages/BreakdownBadgeLocked.svg';
+import oneStaffFaveBadge from './BadgesImages/StaffFave1.svg';
+import oneStaffFaveBadgeLocked from './BadgesImages/StaffFave1Locked.svg';
+import twoStaffFavesBadge from './BadgesImages/StaffFave2.svg';
+import twoStaffFavesBadgeLocked from './BadgesImages/StaffFave2Locked.svg';
+import threeStaffFavesBadge from './BadgesImages/StaffFave3.svg';
+import threeStaffFavesBadgeLocked from './BadgesImages/StaffFave3Locked.svg';
+import voterBadge from './BadgesImages/VoterBadge.svg';
+import voterBadgeLocked from './BadgesImages/VoterBadgeLocked.svg';
+import lock from './BadgesImages/Lock.svg';
+
+const BadgeImages = {
+  signupBadge,
+  signupBadgeLocked,
+  onePostBadge,
+  onePostBadgeLocked,
+  twoPostsBadge,
+  twoPostsBadgeLocked,
+  threePostsBadge,
+  threePostsBadgeLocked,
+  oneStaffFaveBadge,
+  oneStaffFaveBadgeLocked,
+  twoStaffFavesBadge,
+  twoStaffFavesBadgeLocked,
+  threeStaffFavesBadge,
+  threeStaffFavesBadgeLocked,
+  voterBadge,
+  voterBadgeLocked,
+  breakdownBadge,
+  breakdownBadgeLocked,
+  lock,
+};
+
+export default BadgeImages;
