+++ resources/assets/components/pages/AccountPage/Badges/badges-tab.scss
