+++ resources/assets/components/pages/AccountPage/Subscriptions/EmailSubscriptionItem.js
