+++ resources/assets/components/pages/AccountPage/Subscriptions/EmailSubscriptions.js
